refactor : 1차 png -> webp 파일 변환 작업
