+++ src/pages/Write/constants/index.ts
@@ -1,15 +1,15 @@
-import basicTheme from '@/assets/images/basic-theme.png';
-import celebrationStamp from '@/assets/images/celebration-stamp.png';
-import celebration from '@/assets/images/celebration.png';
-import congratTheme from '@/assets/images/congrat-theme.png';
-import consolationStamp from '@/assets/images/consolation-stamp.png';
+import basicTheme from '@/assets/images/basic-theme.webp';
+import celebrationStamp from '@/assets/images/celebration-stamp.webp';
+import celebration from '@/assets/images/celebration.webp';
+import congratTheme from '@/assets/images/congrat-theme.webp';
+import consolationStamp from '@/assets/images/consolation-stamp.webp';
 import consolation from '@/assets/images/consolation.png';
-import consultStamp from '@/assets/images/consult-stamp.png';
-import cunsult from '@/assets/images/consult.png';
+import consultStamp from '@/assets/images/consult-stamp.webp';
+import cunsult from '@/assets/images/consult.webp';
 import etcStamp from '@/assets/images/etc-stamp.png';
 import etc from '@/assets/images/etc.png';
-import fieldTheme from '@/assets/images/field-theme.png';
-import respondentStamp from '@/assets/images/respondent-stamp.png';
+import fieldTheme from '@/assets/images/field-theme.webp';
+import respondentStamp from '@/assets/images/respondent-stamp.webp';
 import skyTheme from '@/assets/images/sky-theme.png';
 import vintageTheme from '@/assets/images/vintage-theme.png';
